[Feat] Modify Hit State and Script
코드 리뷰를 기반으로 힛 상태 와 스크립트 를 수정합니다.
