+++ ItaCH_Smash_Legends/Assets/Script/Player_State/LegendFirstJumpAttackState.cs
@@ -2,8 +2,6 @@ using UnityEngine;
 
 public class LegendFirstJumpAttackState : LegendBaseState
 {
-    // LegnedController 완료시 리펙토링
-
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
@@ -12,9 +10,7 @@ public class LegendFirstJumpAttackState : LegendBaseState
 
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-
         legendController.MoveAndRotate();
         legendController.PlayComboAttack(ComboAttackType.SecondJump);
     }
-
 }
