--- ItaCH_Smash_Legends/Assets/Script/Player_State/LegendHitDownState.cs
@@ -1,13 +1,0 @@
-using UnityEngine;
-
-public class LegendHitDownState : LegendBaseState
-{
-    // LegnedController 완료시 리펙토링
-
-    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    {
-        base.OnStateEnter(animator, stateInfo, layerIndex);
-
-        animator.ResetTrigger(AnimationHash.HitDown);
-    }
-}
