--- ItaCH_Smash_Legends/Assets/Script/Player_State/LegendHitDownState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1afe506ff963c074a924e98da96618ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
