+++ ItaCH_Smash_Legends/Assets/Script/Player_State/LegendHitState.cs
@@ -2,12 +2,10 @@ using UnityEngine;
 
 public class LegendHitState : LegendBaseState
 {
-    // LegnedController 완료시 리펙토링
-
-    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    private EffectController _effectController;
+    public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        base.OnStateEnter(animator, stateInfo, layerIndex);
-
-        animator.ResetTrigger(AnimationHash.Hit);
+        _effectController = animator.GetComponent<EffectController>();
+        _effectController.StartHitFlashEffet().Forget();
     }
 }
