+++ ItaCH_Smash_Legends/Assets/Script/Player_State/LegendHitUpState.cs
@@ -2,29 +2,24 @@ using UnityEngine;
 
 public class LegendHitUpState : LegendBaseState
 {
-    // LegnedController 완료시 리펙토링
-
-    private PlayerHit _playerHit;
-    private PlayerStatus _playerStatus;
+    private EffectController _effectController;
 
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
-
-        _playerHit = animator.GetComponent<PlayerHit>();
-        _playerStatus = animator.GetComponent<PlayerStatus>();
-        _playerStatus.CurrentState = PlayerStatus.State.HitUp;
+        _effectController = animator.GetComponent<EffectController>();
+        _effectController.StartHitFlashEffet().Forget();
     }
-    // 추후 수정
-    //public override void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    //{
-    //    if(animator.GetCurrentAnimatorStateInfo(0).normalizedTime >= 0.6)
-    //    {
-    //    }
-    //}
-    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+
+    public override void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        _playerStatus.CurrentState = PlayerStatus.State.Idle;
-        animator.ResetTrigger(AnimationHash.HitUp);
+        if (legendController.IsFallingOnHitUp())
+        {
+            if (legendController.IsTriggered(ActionType.Jump))
+            {
+                legendController.ResetVelocity();
+                animator.Play(AnimationHash.Jump);
+            }
+        }
     }
 }
